requirements.txt
boto3==1.9.115
dj-database-url==0.5.0
dj-rest-auth==1.1.0
Django==2.1.5
django-allauth==0.40.0
django-daterange-filter==1.3.0
django-cors-headers==3.3.0
django-environ==0.4.5
django-filter==2.2.0
django-forms-bootstrap==3.1.0
django-materializecss-form==1.1.12
django-registration==3.0.1
django-rest-auth==0.9.5
django-storages==1.7.1
django-webpack-loader==0.6.0
djangorestframework==3.9.1

gunicorn==19.9.0
pandas==0.24.1
Pillow==7.0.0
psycopg2-binary==2.8.4
python-dateutil==2.8.0
uk-postcode-utils==1.1